Fixes to Contacts
